requirements.txt
# Core
flask==3.0.3  # Downgraded slightly if issues, but 3.0+ fine
werkzeug==3.0.4
jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
blinker==1.8.2

# AI and data
openai==1.51.2  # Update to latest for better chat API
requests==2.32.3
pandas==2.2.3
pydantic==2.9.2

# Checks and tests
mypy==1.12.0
pylint==3.3.1  # Newer than 4? Wait, pylint 3.3 is latest; astroid accordingly
astroid==3.3.5
isort==5.13.2
mccabe==0.7.0
tomlkit==0.13.2
platformdirs==4.3.6
pathspec==0.12.1
typing_extensions==4.12.2
pytest==8.3.3
dill==0.3.9
colorama==0.4.6  # For Windows, optional

# Utils
python-dotenv==1.0.1  # For .env loading